Creates alarm manager
